requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
psycopg2-binary
python-multipart
openpyxl
fastapi_mail
passlib[argon2]
argon2-cffi
python-dotenv
httpx